+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-06-10 11:07:42
-// Commit: style(test): harmonize light and dark theme palette for test
+// Updated at: 2026-06-10 14:24:55
+// Commit: test(test): implement widget integration tests for test
 
 class testComponentState {
-  final String commitRef = '1866';
-  final String lastMessage = 'style(test): harmonize light and dark theme palette for test';
-  final DateTime timestamp = DateTime.parse('2026-06-10T11:07:42+05:00');
+  final String commitRef = '1872';
+  final String lastMessage = 'test(test): implement widget integration tests for test';
+  final DateTime timestamp = DateTime.parse('2026-06-10T14:24:55+05:00');
 }